requirements.txt
PyGithub
jira
tenacity
click

pytest